requirements.txt
python-musicpd
